refactor: moved xsd_cast_literal function
